Updated app to use prism - Added Prism in DevAssessment project - Updated App.xaml, App.xaml.cs and App.xaml.g.cs to use Prism - Created Views folder - Moved MainPage in Views folder fixes #3
